backend/requirements.txt
fastapi
uvicorn
yt-dlp
openai-whisper
groq
